+++ modules/microapps/client/config/articles.client.routes.js
@@ -37,10 +37,10 @@
       });
   }
 
-  getArticle.$inject = ['$stateParams', 'ArticlesService'];
+  getArticle.$inject = ['$stateParams', 'MicroAppsService'];
 
-  function getArticle($stateParams, ArticlesService) {
-    return ArticlesService.get({
+  function getArticle($stateParams, MicroAppsService) {
+    return MicroAppsService.get({
       articleId: $stateParams.articleId
     }).$promise;
   }