+++ modules/microapps/tests/client/articles.client.controller.tests.js
@@ -8,7 +8,7 @@
       $httpBackend,
       $state,
       Authentication,
-      ArticlesService,
+      MicroAppsService,
       mockArticle;
 
     // The $resource service augments the response object with methods for updating and deleting the resource.
@@ -44,10 +44,10 @@
       $httpBackend = _$httpBackend_;
       $state = _$state_;
       Authentication = _Authentication_;
-      ArticlesService = _ArticlesService_;
+      MicroAppsService = _ArticlesService_;
 
       // create mock article
-      mockArticle = new ArticlesService({
+      mockArticle = new MicroAppsService({
         _id: '525a8422f6d0f87f0e407a33',
         title: 'An Article about MEAN',
         content: 'MEAN rocks!'